tapchannel: refactor result type
This commit unifies the use of the lfn.Result return type as a
preparation for the final commit in this PR, where we change a bunch of
function/method signatures to use the Result type as well.

We use lnf.Err(fnt.Errorf()) instead of lnf.Errf because the
intellisense of some IDEs complain about the '%w' verb only being usable
with fmt.Errorf().