internal/ecies: add encrypt/decrypt with ECIES
This commit creates a simple encryption and decryption function that
uses AES256-GCM for tne symmetric encryption and HKDF with SHA256 for
the key derivation.
The shared key generation is not part of these functions, because we'll
need to use lnd RPCs in some cases to be able to derive it.