universe/supplycommit: implement state transitions for supply commit state machine
In this commit, we build on the two prior commits and add defined state
transitions for the supply commit state machine. We start in a default
state, until we hear about a new update. We'll continue in that state
untli we get a commit tick, at which point we'll enter the main logic of
the state machine.

After a tick, we'll construct the final tree, make a commit transaction
from that, sign it, then write everything to disk. We'll then resume by
broadcasting the commit tx. Once we reecive a confirmation, we'll apply
the state tranition which will update all the relevant state on disk.