tapfreighter: fix bug with burns not being recognized
This fixes the main part of the flake that caused burns to sometimes not
be recognized correctly.
Because we now also use the allocation code to create virtual packets,
we need to apply the same ordering of inputs to the selected commitments
before choosing the first input that the burn key is derived from.
Otherwise it can happen that the actual first input in the allocated
virtual transaction is another one and the resulting burn key isn't
recognized correctly.